requirements.txt
streamlit
groq
python-dotenv
speechrecognition
pyaudio
langdetect
